+++ packages/harpguru-core/src/components/harp-guru-page/harp-guru-page.tsx
@@ -97,9 +97,9 @@ export const HarpGuruPage = ({
             pageOnDisplay.setValue(nextPageNumberMap[thisPage])
           }
         />
+        <ToggleBufferFlusher />
         <QuizQuestionDisplay isScreenFree={menuState === MenuStates.NoMenu} />
         <ScaleNotification />
-        <ToggleBufferFlusher />
       </View>
     </PanGestureHandler>
   )