+++ packages/harpguru-core/src/components/quiz-question-display/utils/index.ts
@@ -1,2 +1,4 @@
 export { activateHarpCell } from './activate-harp-cell'
+export { getCounterpartDegreeId } from './get-counterpart-degree-id'
 export { getNextQuizQuestion } from './get-next-quiz-question'
+export { hasToggledIncorrectCell } from './has-toggled-incorrect-cell'